cron job to read weather information of tasks a day before via weather forecase api

+++ db/migrate/20230730154456_create_user_location_info.rb
@@ -2,8 +2,8 @@ class CreateUserLocationInfo < ActiveRecord::Migration[7.0]
   def change
     create_table :user_location_info do |t|
       t.belongs_to :user, index: { unique: true }, foreign_key: true
-      t.integer :latitude
-      t.integer :longitude
+      t.decimal :latitude
+      t.decimal :longitude
       t.timestamps
     end
   end
